Created a config file for managing paths used in code. Should enable a bit more universal access for people across file systems. May look at moving this file into a different package and we may need to reconsider if this package is ever going public.
database_config.py was added to set up the config file and give DJ paths for root file locations. Subject to change as I use DJ more and realize what it actually needs longterm. This file may get scrapped.

Edited elements file to be more in line with the standard format DJ has used in the past with their pipleine files. Currently, I don't see the benefit of adding in an acticate function since calling that function in other modules will not actually activate anything as they get trapped in local namespace of the function